+++ RAG/Beary/Beary/ValueTypes/Vector.cs
@@ -2,6 +2,6 @@
 
 namespace Beary.ValueTypes;
 
-public class Vector: ValueOf<IEnumerable<Double>, Vector>
+public class Vector: ValueOf<IEnumerable<Single>, Vector>
 {
 }
